requirements.txt
absl-py==2.1.0
astunparse==1.6.3
blinker==1.8.2
build==1.2.1
CacheControl==0.13.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chex==0.1.86
cleo==2.1.0
click==8.1.7
contourpy==1.3.0
crashtest==0.4.1
cryptography==43.0.3
cycler==0.12.1
distlib==0.3.8
dlib==19.24.5
dulwich==0.21.7
etils==1.9.2
face-recognition==1.3.0
face-recognition-models==0.3.0
facenet-pytorch==2.6.0
fastjsonschema==2.19.1
filelock==3.14.0
Flask==3.0.3
flatbuffers==24.3.25
flax==0.8.5
fonttools==4.54.1
fsspec==2024.6.1
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
gprof2dot==2022.7.29
grpcio==1.65.4
h5py==3.11.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
iniconfig==2.0.0
installer==0.7.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jax==0.4.31
jaxlib==0.4.31
Jinja2==3.1.4
keras==3.4.1
keyring==24.3.1
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.3.2
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
opencv-python==4.10.0.84
opt-einsum==3.3.0
optax==0.2.3
optree==0.12.1
orbax-checkpoint==0.5.23
packaging==23.2
pandas==2.2.2
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pexpect==4.9.0
pillow==11.0.0
pkginfo==1.10.0
platformdirs==3.11.0
pluggy==1.5.0
poetry==1.7.0
poetry-core==1.8.1
poetry-dotenv-plugin==0.1.0
poetry-plugin-export==1.6.0
protobuf==4.25.4
psutil==6.1.0
ptyprocess==0.7.0
py-cpuinfo==9.0.0
pycparser==2.22
Pygments==2.18.0
pyparsing==3.2.0
pypdfium2==4.30.0
pyproject_hooks==1.1.0
pytesseract==0.3.13
pytest==8.2.2
pytest-profiling==1.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
rapidfuzz==3.9.1
requests==2.32.2
requests-toolbelt==1.0.0
rich==13.7.1
scipy==1.14.0
seaborn==0.13.2
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sympy==1.13.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.1
tensorflow_decision_forests==1.9.0
tensorflowjs==4.20.0
tensorstore==0.1.64
termcolor==2.4.0
tf_keras==2.16.0
thop==0.1.1.post2209072238
tomlkit==0.12.5
toolz==0.12.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.5
trove-classifiers==2024.5.17
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.3.23
ultralytics-thop==2.0.9
urllib3==2.2.1
virtualenv==20.26.2
Werkzeug==3.0.3
wrapt==1.16.0
wurlitzer==3.1.1
xattr==0.10.1
zipp==3.18.2
